fix(onboard): extract host.docker.internal probe policy
