+++ docs/reference/troubleshooting.mdx
@@ -1073,9 +1073,9 @@ option configures automatically:
 http://host.openshell.internal:11435/v1
 ```
 
-`host.openshell.internal` resolves to the same gateway IP, but port `11435`
-is bound by the [token-gated Ollama auth proxy](#ollama-auth-proxy-did-not-start)
-that forwards requests to `127.0.0.1:11434` on the host.
+`host.openshell.internal` resolves to the same gateway IP, and the
+[token-gated Ollama auth proxy](#ollama-auth-proxy-did-not-start) binds port
+`11435` there and forwards requests to `127.0.0.1:11434` on the host.
 If you need a different host service exposed to the sandbox, route it through
 the OpenShell gateway rather than relying on `host.docker.internal`.
 See issue [#3136](https://github.com/NVIDIA/NemoClaw/issues/3136).
