+++ src/lib/inference/onboard-probes.ts
@@ -9,6 +9,10 @@ const { getCredential, normalizeCredentialValue, resolveProviderCredential } = r
 const { isWsl } = require("../platform");
 const httpProbe = require("../adapters/http/probe");
 const {
+  getHostDockerInternalProbeFailure,
+  isHijackedDockerInternalUrl,
+} = require("./onboard-host-docker-internal");
+const {
   isNvcfFunctionNotFoundForAccount,
   nvcfFunctionNotFoundMessage,
   shouldForceCompletionsApi,
@@ -30,24 +34,10 @@ const {
 // must fail closed unless the host is probeable from the onboard process.
 const SANDBOX_INTERNAL_HOSTS = ["host.openshell.internal"];
 
-// host.docker.internal is not a stable host-service route from OpenShell k3s
-// sandboxes. Depending on the platform it may fail DNS resolution or resolve to
-// a gateway/bridge address where the user's host service is not forwarded. See
-// issue #3136. Always steer users to a NemoClaw-managed proxy URL instead.
-const HOST_DOCKER_INTERNAL = "host.docker.internal";
-
 function isSandboxInternalUrl(url) {
   try {
     const { hostname } = new URL(String(url));
     return SANDBOX_INTERNAL_HOSTS.includes(hostname);
-  } catch {
-    return false;
-  }
-}
-
-function isHijackedDockerInternalUrl(url) {
-  try {
-    return new URL(String(url)).hostname === HOST_DOCKER_INTERNAL;
   } catch {
     return false;
   }
@@ -499,27 +489,7 @@ function runChatCompletionsProbe({ authHeader, model, url, isWsl: isWslOverride 
 
 function probeOpenAiLikeEndpoint(endpointUrl, model, apiKey, options = {}) {
   if (isHijackedDockerInternalUrl(endpointUrl) && options.allowHostDockerInternal !== true) {
-    return {
-      ok: false,
-      message:
-        `${HOST_DOCKER_INTERNAL} does not reach the host machine from inside the sandbox: ` +
-        `OpenShell k3s sandboxes do not provide it as a reliable host-service route. ` +
-        `It may fail DNS resolution or resolve to a gateway/bridge address where port ` +
-        `11434 is not forwarded. For local Ollama, use the auth-proxy URL ` +
-        `http://host.openshell.internal:11435/v1 (the URL NemoClaw onboard configures ` +
-        `automatically when you pick "Local Ollama"). See issue #3136.`,
-      failures: [
-        {
-          name: "host.docker.internal reachability",
-          httpStatus: 0,
-          curlStatus: 0,
-          message:
-            `${HOST_DOCKER_INTERNAL} is not a reliable host-service route from ` +
-            `OpenShell k3s sandboxes and cannot be used as an inference base URL.`,
-          body: "",
-        },
-      ],
-    };
+    return getHostDockerInternalProbeFailure();
   }
 
   if (isSandboxInternalUrl(endpointUrl)) {
